requirements.txt
# MySQL JDBC connector
mysql-connector-python==8.0.33

# YAML configuration
PyYAML==6.0.1